Cuarto punto dev

+++ punto1.py
@@ -5,7 +5,7 @@ Created on Wed Mar  2 19:47:25 2022
 @author: dbenavid5
 """
 
-valorDonacion = int(input('Ingrese el valor de la donacion: \n'))
+valorDonacion = float(input('Ingrese el valor de la donacion: \n'))
 administracion = valorDonacion * 0.35;
 sistemas = administracion * 0.25;
 telecomunicaciones = sistemas *0.1;
