+++ punto2.py
@@ -5,9 +5,9 @@ Created on Wed Mar  2 19:51:29 2022
 @author: dbenavid5
 """
 
-inversion1 = int(input('Ingrese el valor de la primera inversion: \n'))
-inversion2 = int(input('Ingrese el valor de la segunda inversion: \n'))
-inversion3 = int(input('Ingrese el valor de la tercera inversion: \n'))
+inversion1 = float(input('Ingrese el valor de la primera inversion: \n'))
+inversion2 = float(input('Ingrese el valor de la segunda inversion: \n'))
+inversion3 = float(input('Ingrese el valor de la tercera inversion: \n'))
 valorTotalInv = inversion1 + inversion2 + inversion3;
 print(f"El valor total de la inversion es: ${valorTotalInv:,}\n")
 pi1 = (inversion1/valorTotalInv) * 100
